spring security

+++ src/main/java/com/weedoctor/api/repo/UserJpaRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.yuequan.jpa.soft.delete.repository.SoftDelete;
 
 @SoftDelete
-public interface UserJpaRepo extends JpaRepository<Users, Integer> {
+public interface UserJpaRepo extends JpaRepository<Users, Long> {
 
 }
